requirements.txt
matplotlib==3.6.2
numpy==1.23.1
flwr==1.4.0
pytorch==1.12.1
torchvision==0.13.1
pandas==2.1.2
transformers==4.36.0
